+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Refactor context-building utilities into ctyutil (for possible later moving to an external repo) and hclutil
 - Reorganize code and move to plugin registries for functions, servers, clients, triggers, and ambient values (env.*, httpstatus.*, sys.*)
 
 ## [0.18.0] - 2026-03-28